putnbr big values

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: atastet <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/10 15:45:49 by atastet           #+#    #+#             */
-/*   Updated: 2018/04/24 15:05:08 by atastet          ###   ########.fr       */
+/*   Updated: 2018/04/24 17:55:21 by atastet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_atoi(const char *str)
 {
-	int res;
+	int	res;
 	int i;
-	int sign;
+	int	sign;
 
 	i = 0;
 	res = 0;
@@ -35,5 +35,5 @@ int	ft_atoi(const char *str)
 		res = res * 10 + (str[i] - '0');
 		i++;
 	}
-	return (res * sign);
+	return ((int)(res * sign));
 }
